Implement a basic battle loop (#3)
* Create basic skeleton for Battle Flow

* Start battle through debug UI

* Select skill, then display target

In battle-flow-test.unity, hit Play.
Click on the 'Start Battle' button in the Game Window to display a list
of skills for the first Combatant.
Click a skill to display the possible targets for that skill.

* Remove Start and Update functions

* Selecting target executes Action

* Add damage to Bash/Slash/Stab ATK

* Actually execute action

* Loop combat seamlessly

Once an Action is executed, it will begin the Action building process
for the next Combatant.
This continues over and over the sorted queue of Combatants.

Still requires some sort of 'exit', and cleanly checking for 'Can
Combat Continue?'

* Refactor battle state graph, CanContinueBattle()

Consolidated two successive nodes in the battle state graph. Also, we
now actually check if the battle can continue by checking:
1. Are we in battle right now?
2. Are the heroes alive?
3. Are the enemies alive?

If any are false, we exit battle.
Test by casting SLASH as either the Goblin Hero or its ally. This will
instantly kill the Goblin enemies and end the fight.