Handle list of null skills, fix combatantdata
For some reason, the skills associated with the Goblin Monk, Human
Knight, and Slime were unlinked. This manifested as the Goblin Monk
having 3 null skills and etc. This commit should fix the unlinking as
well as handle future unlinking situations more gracefully in
SkillButtonLister.

Also updates the InputManager.asset with Start, Square, and Triangle