Start using new Input System in Bolt
The first transition out of 'Out of Battle' in battle_state now uses a
custom Bolt SuperUnit that wraps around the new Unity Input system. Just
have to hook it up to BattleTester to actually start it.

Also, it seems to trigger twice; on the press of the Options button on
the DS4 and on the release. Should make a SuperUnit for each thing. I
feel this will also have complications if I have some input that is
active while a button is held down.