Fix broken ActionTest
With the conversion of Skill.Execute to an IEnumerator, the assertions
of ActionTest had to be passed into the Action/Skill.Execute function as
a delegate.

Future me, heed these warnings well: watch your callback function use!
Be clear if a function was only designed to run concurrently--it may be run
synchronously!