Use BattleManager in comb. AI creation, gameover
In the battle_state flow graph, I was using the ActionBuilderUI to get
the current combatant to randomly choose an Action. I am now getting it
from BattleManager directly.

Still have to make a 'game over' state, as well as fix the bug that
causes the AI to choose an already-dead target.